requirements.txt
MouseInfo==0.1.3
numpy==1.19.4
opengles==0.0.1
Pillow==8.0.1
PyAutoGUI==0.9.52
pygame==2.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python3-xlib==0.15
PyTweening==1.0.3
